pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "swarm-mcp"
version = "3.4.0"
description = "Algorithmic Agentic Orchestrator with HippoRAG and Autonomous Git Worker"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Swarm Team", email = "team@swarm-mcp.dev"}
]
dependencies = [
    "pydantic>=2.9.0",
    "typer>=0.12.0",
    "filelock>=3.15.0",
    "rich>=13.9.0",
    "fastmcp>=2.0.0",
    "google-generativeai>=0.3.0",
    "openai>=1.0.0",
    "httpx>=0.27.0",

    "networkx>=3.0",
    "coverage>=7.0",
    "pytest>=8.3.0",
    "flake8>=7.1.0",
    "asyncpg>=0.29.0",
    "pgvector>=0.2.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0"
]

[project.optional-dependencies]
treesitter = [
    "tree-sitter>=0.23.0",
    "tree-sitter-javascript>=0.23.0",
    "tree-sitter-typescript>=0.23.0"
]
verification = [
    "z3-solver>=4.12.0"
]
dev = [
    "pytest-cov",
    "anyio",
    "pytest-asyncio",
    "mutmut>=2.4.0",
    "black",
    "isort"
]

[project.scripts]
swarm = "orchestrator:app"

[tool.setuptools]
py-modules = ["orchestrator", "server", "dashboard_server"]

[tool.setuptools.packages.find]
where = ["."]
include = ["mcp_core*", "scripts*"]
